3834 - Moving trace building related files for avm into vm folder and introducing new namespace
