+++ barretenberg/cpp/src/barretenberg/vm/avm_trace/AvmMini_helper.hpp
@@ -2,8 +2,8 @@
 
 #include "AvmMini_common.hpp"
 
-namespace proof_system {
+namespace avm_trace {
 
 void log_avmMini_trace(std::vector<Row> const& trace, size_t beg, size_t end);
 
-} // namespace proof_system
+} // namespace avm_trace
