+++ barretenberg/cpp/src/barretenberg/vm/avm_trace/AvmMini_mem_trace.cpp
@@ -1,6 +1,6 @@
 #include "AvmMini_mem_trace.hpp"
 
-namespace proof_system {
+namespace avm_trace {
 
 /**
  * @brief Constructor of a memory trace builder of AVM. Only serves to set the capacity of the
@@ -241,4 +241,4 @@ void AvmMiniMemTraceBuilder::write_into_memory(
     store_in_mem_trace(clk, interm_reg, addr, val, m_in_tag);
 }
 
-} // namespace proof_system
+} // namespace avm_trace
